rag_indexer/requirements.txt
functions-framework==3.*
vertexai
google-cloud-firestore
google-cloud-pubsub
